+++ src/tg2go/db/repositories/order.py
@@ -1,5 +1,5 @@
 import logging
-
+    
 from sqlalchemy.ext.asyncio import AsyncSession, async_sessionmaker
 
 from tg2go.db.models.good import Good, GoodId
@@ -11,6 +11,7 @@ class OrderRepository:
     def __init__(self, session: async_sessionmaker[AsyncSession]):
         self.session = session
 
+    # ----- Create -----
     async def CreateNewOrder(self, chat_id: int) -> OrderId:
         order = CreateOrder(chat_id=chat_id)
 
@@ -22,6 +23,7 @@ class OrderRepository:
 
         return order.order_id
 
+    # ----- Update -----
     async def AddGoodToOrder(self, order_id: OrderId, good_id: GoodId) -> None:
         async with self.session() as session:
             order = await session.get(Order, order_id)
@@ -72,15 +74,17 @@ class OrderRepository:
             for item in order.order_items:
                 if good.good_id == item.good_id:
                     item.quantity -= 1
-
-                    if item.quantity == 0:
-                        await session.delete(item)
-
                     break
             else:
                 raise ValueError(f"No {good} in {order} when removing good from order.")
 
             order.total_price_rub -= item.unit_price_rub
+
+            if item.quantity == 0:
+                await session.delete(item)
+
+            if order.total_price_rub < 0:
+                raise ValueError(f"Total price RUB is negative for {order}")
 
             logging.info(f"After removing good {good} order is {order}")